Git: ensure LFS works together with noshared
"noshared" mode is designed so the build can be run eg. in a different
filesystem namespace, with git objects being still available: git-clone
without "-s" makes a copy of them.

However, this is not sufficient for git-lfs, whose objects are not
copied using git-clone.  This copies (when needed) the contents of the
LFS cache from DL_DIR into the cache of the just-cloned worktree.

Note this is over-eager, and may copy more data than necessary as the
DL_DIR ages.  Being more selective would require identifying the lfs
objects about to be requested, before doing the checkout.

Signed-off-by: Yann Dirson <yann.dirson@vates.tech>